+++ lib/punchblock.rb
@@ -40,6 +40,21 @@ module Punchblock
     def reset_logger
       @logger = NullObject.new
     end
+
+    #
+    # Get a new Punchblock client with a connection attached
+    #
+    # @param [Symbol] type the connection type (eg :XMPP, :asterisk)
+    # @param [Hash] options the options to pass to the connection (credentials, etc
+    #
+    # @return [Punchblock::Client] a punchblock client object
+    #
+    def client_with_connection(type, options)
+      connection = Connection.const_get(type.to_s.classify).new options
+      Client.new :connection => connection
+    rescue NameError
+      raise ArgumentError, "Connection type #{type.inspect} is not valid."
+    end
   end
 
   ##