+++ lib/punchblock/translator/asterisk.rb
@@ -19,6 +19,8 @@ module Punchblock
       REDIRECT_EXTENSION = '1'
       REDIRECT_PRIORITY = '1'
 
+      trap_exit :actor_died
+
       def initialize(ami_client, connection, media_engine = nil)
         pb_logger.debug "Starting up..."
         @ami_client, @connection, @media_engine = ami_client, connection, media_engine
@@ -32,8 +34,8 @@ module Punchblock
       end
 
       def deregister_call(call)
-        @channel_to_call_id[call.channel] = nil
-        @calls[call.id] = nil
+        @channel_to_call_id.delete call.channel
+        @calls.delete call.id
       end
 
       def call_with_id(call_id)
@@ -107,7 +109,7 @@ module Punchblock
         if call = call_with_id(command.target_call_id)
           call.execute_command! command
         else
-          command.response = ProtocolError.new.setup 'item-not-found', "Could not find a call with ID #{command.target_call_id}", command.target_call_id
+          command.response = ProtocolError.new.setup :item_not_found, "Could not find a call with ID #{command.target_call_id}", command.target_call_id
         end
       end
 
@@ -115,7 +117,7 @@ module Punchblock
         if (component = component_with_id(command.component_id))
           component.execute_command! command
         else
-          command.response = ProtocolError.new.setup 'item-not-found', "Could not find a component with ID #{command.component_id}", command.target_call_id, command.component_id
+          command.response = ProtocolError.new.setup :item_not_found, "Could not find a component with ID #{command.component_id}", command.target_call_id, command.component_id
         end
       end
 
@@ -126,7 +128,7 @@ module Punchblock
           register_component component
           component.execute!
         when Punchblock::Command::Dial
-          call = Call.new command.to, current_actor
+          call = Call.new_link command.to, current_actor
           register_call call
           call.dial! command
         else
@@ -143,6 +145,18 @@ module Punchblock
           'Command' => "dialplan add extension #{REDIRECT_EXTENSION},#{REDIRECT_PRIORITY},AGI,agi:async into #{REDIRECT_CONTEXT}"
         })
         pb_logger.trace "Added extension extension #{REDIRECT_EXTENSION},#{REDIRECT_PRIORITY},AGI,agi:async into #{REDIRECT_CONTEXT}"
+      end
+
+      def actor_died(actor, reason)
+        return unless reason
+        pb_logger.error "A linked actor (#{actor.inspect}) died due to #{reason.inspect}"
+        if id = @calls.key(actor)
+          pb_logger.info "Dead actor was a call we know about, with ID #{id}. Removing it from the registry..."
+          @calls.delete id
+          end_event = Punchblock::Event::End.new :target_call_id  => id,
+                                                 :reason          => :error
+          handle_pb_event end_event
+        end
       end
 
       private
@@ -186,6 +200,7 @@ module Punchblock
 
         pb_logger.trace "Handling AsyncAGI Start event by creating a new call"
         call = Call.new event['Channel'], current_actor, env
+        link call
         register_call call
         call.send_offer!
       end